Изменил структуру проекта + сообщения снова обновляются + теперь всё стабильно
