+++ .gitignore
@@ -1,0 +1,5 @@
+test.py
+*.session
+*.session-journal
+__pycache__
+*/__pycache__
